Added definition for the static member alloc
C++ Primer 5th edition errata: Pages 525–528: The alloc member of StrVec should be a static member of the class, not a data member of each object. To fix this error, add the static keyword to the declaration of alloc inside the StrVec class and provide a definition of the form
                                allocator<string> StrVec::alloc;
inside the implementation file for StrVec.