Update main.cpp
Dynamic binding is definitely happening when you call debug() through a reference or pointer to the base class Quote. The bug in your code (the first part before std::cout << "====================\n";) is that once initialized, a reference cannot be changed to refer to another object. So the reference "r" is always referring to the base class object q (although you assign bq and lq to it later) and that's why r.debug() is always the base class version. Fix it by using pointer rather than reference. Also removing the misleading comments.